Fix unit tests for Chef.
Unit test is set up to close resources at the end of the test.

This still doesn't feel as intuitive as it can--it's implementing
Closeable, which, while it's a powerful design pattern, is still a Java
interface. Since we're trying to make this as far Kotlin in one
direction as we can, I want to try to find a more idiomatically Kotlin
way to clean up after the Chefs.

Perhaps the restaurant itself manages the jobs. Or a literal
RestaurantManager job that closes other jobs (but who closes their
job?).

In any case, it seems like a Chef should implement a Job, so it's much
easier to imagine how things interact with each other, as it enforces
the role-playing and the metaphor.