Remove TODO in Chef.kt.
After deliberating on whether or not Chef should implement Job, it now
seems like it shouldn't be the case where it should delegate to an inner
Job element. It seems like using launch and async is important for
understanding that some form of coroutine management should be
considered.

A more object-oriented approach may mean that objects are responsible
for canceling the resources they are involved in. The resource
management (managing Chefs) may be more interesting as the Waiter and
Customer classes are installed with coroutines.