Removed all unused code.
When I worked on this last week, I had no idea what was happening and I
was getting fairly overwhelmed with learning all of the different tools
to build this Restaurant. I'm going to focus on one coroutine/Job/aspect
at a time and slowly learn my way through this with lots of
documentation and unit/learning tests.