Create Customer and Waiter class.
The customer class has a very simple delay functionality. The complexity
in terms of coroutine launching will be on the burden of the Waiter to
start the decision and eating Jobs for the Customer, so minimal
functionality was written to the Waiter class.

For now, each Waiter can serve only 1 customer and manage that
Customer's pipeline (their dining experience). After implementing this
sequence, I'll empower our Waiter to take n Customers, and test Waiters
against super-Waiters that can take serve lots of Customers at once (or
naive, ambitious Waiters that say they can take a lot at once, but are
very slow).