Implement most of Waiter's waiting algorithm.
The waiter will be periodically checking on the customer given their own
checkInDelay parameter. This allows us to simulate delays of which we
have no control over--say some sort of network call or any external
connection.

Perhaps it's even more succintly simulated in the yet-to-be-injected
suspend function for the Waiter--sendOrder(Order) -> Meal is a
suspending function that allows the Waiter to simply wait for that order
to cook regardless of the wait time. A waiter can then delegate this
behavior into a single function and simply wait for that function by
suspending itself. It seems like as long as I believe that suspending a
coroutine is extremely cheap relative to how expensive it is to block an
entire thread, I should be able to suspend and wait for things to no
end and we can have suspend functions all the way down.

Anyway, starting to see the power of coroutines and suspension functions
now. More to come when I actually unit test this class.